composer: change requirement to support 8.1
Wikimedia Foundation is still relying on PHP 8.1 for production and we
try to keep supporting it when feasible.

* lower PHPUnit requirement from 12.x to 10.x
* wegenerate composer.lock with `composer update -w phpunit/phpunit`
* in composer, instead of config.platform.php, use require.php:

Composer config.platform.php=8.3 causes composer to resolve dependencies
as if it ran under PHP 8.3.0. It is not necessarily correct as described
on https://getcomposer.org/doc/06-config.md#platform

Instead, use the recommended `require.php` and set it to `php8>=8.1.0`

Change-Id: I13d6c07029e3a3f280df78f838551274fe29a256